Delete .github/workflows/rust.yml
Well, that didn't work so well :)

Biggest problem I think is that Github probably wouldn't like running an assembler and linker (i.e. two basically random exe files) on their servers, but more research can be done on this later to see if this is actually an issue or not.

The other problem is that this program needs to run on Windows, not Ubuntu, which is the default provided by Github. Again, maybe this can be fixed.