Input bug fixed
When users entered input, it was suffixed with a carriage return and newline character - this input was then immediately fed to functions like string_to_float which expected the input to be null terminated. This bug is now fixed, as all input into the program is sanitized and null terminated, with no carriage return or newline characters slipping past.